sdk-sample-scripts/Solarwinds/requirements.txt
flatten_json>=0.1.14
requests>=2.32.3
runzero-sdk>=0.8.2
orionsdk>=0.4.0